moved GlobalSettings to common and created PR
