--- msal/src/main/res/raw/msal_default_global_config.json
@@ -1,117 +1,0 @@
-{
-  "minimum_required_broker_protocol_version": "3.0",
-  "web_view_zoom_controls_enabled": true,
-  "web_view_zoom_enabled": true,
-  "power_opt_check_for_network_req_enabled": true,
-  "handle_null_taskaffinity": false,
-  "authorization_in_current_task": false,
-  "logging": {
-    "pii_enabled": false,
-    "log_level": "WARNING",
-    "logcat_enabled": true
-  },
-  "browser_safelist": [
-    {
-      "browser_package_name": "com.android.chrome",
-      "browser_signature_hashes": [
-        "7fmduHKTdHHrlMvldlEqAIlSfii1tl35bxj1OXN5Ve8c4lU6URVu4xtSHc3BVZxS6WWJnxMDhIfQN0N0K2NDJg=="
-      ]
-    },
-    {
-      "browser_package_name": "org.mozilla.firefox",
-      "browser_signature_hashes": [
-        "2gCe6pR_AO_Q2Vu8Iep-4AsiKNnUHQxu0FaDHO_qa178GByKybdT_BuE8_dYk99G5Uvx_gdONXAOO2EaXidpVQ=="
-      ]
-    },
-    {
-      "browser_package_name": "com.sec.android.app.sbrowser",
-      "browser_signature_hashes": [
-        "ABi2fbt8vkzj7SJ8aD5jc4xJFTDFntdkMrYXL3itsvqY1QIw-dZozdop5rgKNxjbrQAd5nntAGpgh9w84O1Xgg=="
-      ]
-    },
-    {
-      "browser_package_name": "com.cloudmosa.puffinFree",
-      "browser_signature_hashes": [
-        "1WqG8SoK2WvE4NTYgr2550TRhjhxT-7DWxu6C_o6GrOLK6xzG67Hq7GCGDjkAFRCOChlo2XUUglLRAYu3Mn8Ag=="
-      ]
-    },
-    {
-      "browser_package_name": "com.duckduckgo.mobile.android",
-      "browser_signature_hashes": [
-        "S5Av4cfEycCvIvKPpKGjyCuAE5gZ8y60-knFfGkAEIZWPr9lU5kA7iOAlSZxaJei08s0ruDvuEzFYlmH-jAi4Q=="
-      ]
-    },
-    {
-      "browser_package_name": "com.explore.web.browser",
-      "browser_signature_hashes": [
-        "BzDzBVSAwah8f_A0MYJCPOkt0eb7WcIEw6Udn7VLcizjoU3wxAzVisCm6bW7uTs4WpMfBEJYf0nDgzTYvYHCag=="
-      ]
-    },
-    {
-      "browser_package_name": "com.ksmobile.cb",
-      "browser_signature_hashes": [
-        "lFDYx1Rwc7_XUn4KlfQk2klXLufRyuGHLa3a7rNjqQMkMaxZueQfxukVTvA7yKKp3Md3XUeeDSWGIZcRy7nouw=="
-      ]
-    },
-    {
-      "browser_package_name": "com.microsoft.emmx",
-      "browser_signature_hashes": [
-        "Ivy-Rk6ztai_IudfbyUrSHugzRqAtHWslFvHT0PTvLMsEKLUIgv7ZZbVxygWy_M5mOPpfjZrd3vOx3t-cA6fVQ=="
-      ]
-    },
-    {
-      "browser_package_name": "com.opera.browser",
-      "browser_signature_hashes": [
-        "FIJ3IIeqB7V0qHpRNEpYNkhEGA_eJaf7ntca-Oa_6Feev3UkgnpguTNV31JdAmpEFPGNPo0RHqdlU0k-3jWJWw=="
-      ]
-    },
-    {
-      "browser_package_name": "com.opera.mini.native",
-      "browser_signature_hashes": [
-        "TOTyHs086iGIEdxrX_24aAewTZxV7Wbi6niS2ZrpPhLkjuZPAh1c3NQ_U4Lx1KdgyhQE4BiS36MIfP6LbmmUYQ=="
-      ]
-    },
-    {
-      "browser_package_name": "mobi.mgeek.TunnyBrowser",
-      "browser_signature_hashes": [
-        "RMVoXuK1sfJZuGZ8onG1yhMc-sKiAV2NiB_GZfdNlN8XJ78XEE2wPM6LnQiyltF25GkHiPN2iKQiGwaO2bkyyQ=="
-      ]
-    },
-    {
-      "browser_package_name": "org.mozilla.focus",
-      "browser_signature_hashes": [
-        "L72dT-stFqomSY7sYySrgBJ3VYKbipMZapmUXfTZNqOzN_dekT5wdBACJkpz0C6P0yx5EmZ5IciI93Q0hq0oYA=="
-      ]
-    },
-    {
-      "browser_package_name": "com.cake.browser",
-      "browser_signature_hashes": [
-        "442kvSdZT1fEAewzSi8Wre73x4mWmHBhOFtQ-9T9N6ExZzUdsELUmaaS0edsI7ur2nY-bjbWX7IpluFOyvKkOA=="
-      ]
-    },
-    {
-      "browser_package_name": "com.brave.browser",
-      "browser_signature_hashes": [
-        "wIwX1v_1TfPxHm5qn-_jdGoH3Pa9VVMR5dtVz0Y0xqPkyM_KlavjWPSgOolrVH05AVO1cHWoLPqMzCH04Pw8LQ=="
-      ]
-    },
-    {
-      "browser_package_name": "com.kiwibrowser.browser",
-      "browser_signature_hashes": [
-        "kmPeixKA04JcDuWNBMUPu_6WaODr6a9ofROUIHIGxiiFGvH8Y92MonrDQmsNqEJO2DQkpEQc425WmAYB4NlD3Q=="
-      ]
-    },
-    {
-      "browser_package_name": "com.mi.globalbrowser.mini",
-      "browser_signature_hashes": [
-        "6FEWlPfWn-omfES2ZYDj5bZUIR5au_nfyRr-o_1R3fesjfoV1JptBMumtvVIo0q37abcMRWQt9RUSNXpzKpNdA=="
-      ]
-    },
-    {
-      "browser_package_name": "mark.via.gp",
-      "browser_signature_hashes": [
-        "oTJf5e5nB1NinkdBpkkmhPnwbVRmDlHZ-s_QhvyuGKM5nq5XtjA439O31wxrkL6ReHyyKfDHFUHpQnoXoj--Ig=="
-      ]
-    }
-  ]
-}
